// src/client/slaops-cloud/models/create-service-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SLAOps API
 * Internal API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface CreateServiceDto {
    /**
     * User ID who owns this service
     */
    'user_id': string;
    /**
     * Service name
     */
    'name': string;
    /**
     * Service endpoint URL
     */
    'endpoint': string;
    /**
     * OpenAPI document URL
     */
    'openapi_doc_url'?: string | null;
    /**
     * OpenAPI document content (stored as text)
     */
    'openapi_doc_content'?: string | null;
    /**
     * Service availability percentage
     */
    'availability'?: number | null;
    /**
     * Average response time in milliseconds
     */
    'response_time'?: number | null;
}